requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
azure-search-documents>=11.4.0
azure-identity>=1.15.0
azure-storage-blob
pinecone-client>=3.0.0

typer>=0.9.0
python-dotenv>=1.0.0
